Revert GL Interface Overhaul
Reason: The shared ptr causes too much issue. If user sets a static var as shared ptr, and forgot to clean up, it crashes the program on exit as gl lib uninited before the object itself is unibited. One way to fix this is by keeping track of ALL objects created and delete it on unload.

This reverts commit eb6f6670ea0f82f8e645069a2836bcbb73347b84.