Shader: Text Shader && FlatRect Shader bug fix.
TextRender2.geom:
Fixed the issue where the char is rendered to the left of the char pos instead of right. Cause: left and right flipped.

FlatRect.geom:
Fixed the issue where the texture is rendered 4 times or half the size. Cause: Texture chords range was -1 to 1, where the correct value range is 0 to 1.